fix: some lints and such
