+++ crates/consensus/src/receipt/receipts.rs
@@ -1,0 +1,200 @@
+use alloy_primitives::{Bloom, Log};
+use alloy_rlp::{length_of_length, BufMut, Decodable, Encodable};
+
+/// Receipt containing result of transaction execution.
+#[derive(Clone, Debug, PartialEq, Eq, Default)]
+pub struct Receipt {
+    /// If transaction is executed successfully.
+    ///
+    /// This is the `statusCode`
+    pub success: bool,
+    /// Gas used
+    pub cumulative_gas_used: u64,
+    /// Log send from contracts.
+    pub logs: Vec<Log>,
+}
+
+impl Receipt {
+    /// Calculates [`Log`]'s bloom filter. this is slow operation and [ReceiptWithBloom] can
+    /// be used to cache this value.
+    pub fn bloom_slow(&self) -> Bloom {
+        self.logs.iter().collect()
+    }
+
+    /// Calculates the bloom filter for the receipt and returns the [ReceiptWithBloom] container
+    /// type.
+    pub fn with_bloom(self) -> ReceiptWithBloom {
+        self.into()
+    }
+}
+
+impl alloy_network::Receipt for Receipt {
+    fn success(&self) -> bool {
+        self.success
+    }
+
+    fn bloom(&self) -> Bloom {
+        self.bloom_slow()
+    }
+
+    fn cumulative_gas_used(&self) -> u64 {
+        self.cumulative_gas_used
+    }
+
+    fn logs(&self) -> &[Log] {
+        &self.logs
+    }
+}
+
+/// [`Receipt`] with calculated bloom filter.
+///
+/// This convenience type allows us to lazily calculate the bloom filter for a
+/// receipt, similar to [`Sealed`].
+///
+/// [`Sealed`]: ::alloy_network::Sealed
+#[derive(Clone, Debug, PartialEq, Eq, Default)]
+pub struct ReceiptWithBloom {
+    /// The receipt.
+    pub receipt: Receipt,
+    /// The bloom filter.
+    pub bloom: Bloom,
+}
+
+impl From<Receipt> for ReceiptWithBloom {
+    fn from(receipt: Receipt) -> Self {
+        let bloom = receipt.bloom_slow();
+        ReceiptWithBloom { receipt, bloom }
+    }
+}
+
+impl ReceiptWithBloom {
+    /// Create new [ReceiptWithBloom]
+    pub const fn new(receipt: Receipt, bloom: Bloom) -> Self {
+        Self { receipt, bloom }
+    }
+
+    /// Consume the structure, returning only the receipt
+    #[allow(clippy::missing_const_for_fn)] // false positive
+    pub fn into_receipt(self) -> Receipt {
+        self.receipt
+    }
+
+    /// Consume the structure, returning the receipt and the bloom filter
+    #[allow(clippy::missing_const_for_fn)] // false positive
+    pub fn into_components(self) -> (Receipt, Bloom) {
+        (self.receipt, self.bloom)
+    }
+
+    fn payload_len(&self) -> usize {
+        self.receipt.success.length()
+            + self.receipt.cumulative_gas_used.length()
+            + self.bloom.length()
+            + self.receipt.logs.length()
+    }
+
+    /// Returns the rlp header for the receipt payload.
+    fn receipt_rlp_header(&self) -> alloy_rlp::Header {
+        alloy_rlp::Header { list: true, payload_length: self.payload_len() }
+    }
+
+    /// Encodes the receipt data.
+    fn encode_fields(&self, out: &mut dyn BufMut) {
+        self.receipt_rlp_header().encode(out);
+        self.receipt.success.encode(out);
+        self.receipt.cumulative_gas_used.encode(out);
+        self.bloom.encode(out);
+        self.receipt.logs.encode(out);
+    }
+
+    /// Decodes the receipt payload
+    fn decode_receipt(buf: &mut &[u8]) -> alloy_rlp::Result<Self> {
+        let b: &mut &[u8] = &mut &**buf;
+        let rlp_head = alloy_rlp::Header::decode(b)?;
+        if !rlp_head.list {
+            return Err(alloy_rlp::Error::UnexpectedString);
+        }
+        let started_len = b.len();
+
+        let success = Decodable::decode(b)?;
+        let cumulative_gas_used = Decodable::decode(b)?;
+        let bloom = Decodable::decode(b)?;
+        let logs = Decodable::decode(b)?;
+
+        let receipt = Receipt { success, cumulative_gas_used, logs };
+
+        let this = Self { receipt, bloom };
+        let consumed = started_len - b.len();
+        if consumed != rlp_head.payload_length {
+            return Err(alloy_rlp::Error::ListLengthMismatch {
+                expected: rlp_head.payload_length,
+                got: consumed,
+            });
+        }
+        *buf = *b;
+        Ok(this)
+    }
+}
+
+impl alloy_network::Receipt for ReceiptWithBloom {
+    fn success(&self) -> bool {
+        self.receipt.success
+    }
+
+    fn bloom(&self) -> Bloom {
+        self.bloom
+    }
+
+    fn bloom_cheap(&self) -> Option<Bloom> {
+        Some(self.bloom)
+    }
+
+    fn cumulative_gas_used(&self) -> u64 {
+        self.receipt.cumulative_gas_used
+    }
+
+    fn logs(&self) -> &[Log] {
+        &self.receipt.logs
+    }
+}
+
+impl alloy_rlp::Encodable for ReceiptWithBloom {
+    fn encode(&self, out: &mut dyn BufMut) {
+        self.encode_fields(out);
+    }
+
+    fn length(&self) -> usize {
+        let payload_length = self.receipt.success.length()
+            + self.receipt.cumulative_gas_used.length()
+            + self.bloom.length()
+            + self.receipt.logs.length();
+        payload_length + length_of_length(payload_length)
+    }
+}
+
+impl alloy_rlp::Decodable for ReceiptWithBloom {
+    fn decode(buf: &mut &[u8]) -> alloy_rlp::Result<Self> {
+        Self::decode_receipt(buf)
+    }
+}
+
+#[cfg(any(test, feature = "arbitrary"))]
+impl<'a> arbitrary::Arbitrary<'a> for Receipt {
+    fn arbitrary(u: &mut arbitrary::Unstructured<'a>) -> arbitrary::Result<Self> {
+        let success = bool::arbitrary(u)?;
+        let cumulative_gas_used = u64::arbitrary(u)?;
+        let logs = Vec::<Log>::arbitrary(u)?;
+
+        // Only receipts for deposit transactions may contain a deposit nonce
+        #[cfg(feature = "optimism")]
+        let deposit_nonce =
+            if tx_type == TxType::DEPOSIT { Option::<u64>::arbitrary(u)? } else { None };
+
+        Ok(Self {
+            success,
+            cumulative_gas_used,
+            logs,
+            #[cfg(feature = "optimism")]
+            deposit_nonce,
+        })
+    }
+}
